feat: 게시글, 댓글 삭제 및 수정 시 확인 절차 추가 1. 게시글 삭제 및 수정 시 게시글의 유저 아이디와 현재 세션을 통하여 로그인한 유저의 아이디가 같은 지 확인 2. 불일치 시 failed 리턴, 일치 시 이후 로직 진행 3. 댓글 삭제도 동일하게 진행

+++ src/main/java/com/example/todayrecipe/comment/service/impl/CommentServiceImpl.java
@@ -14,6 +14,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -62,9 +63,18 @@ public class CommentServiceImpl implements CommentService {
     }
     @Transactional
     @Override
-    public String deleteComment(Long comment_id) {
+    public String deleteComment(Long comment_id, String userId) {
         try{
-            repo.deleteCommentById(comment_id);
+            User user = userRepo.findUserByUserid(userId).orElse(null);
+            Comment comment = repo.findCommentById(comment_id);
+            Long user_Id = user.getId();
+            Long commentId = comment.getUser().getId();
+            if (!user_Id.equals(commentId)) {
+                return "failed";
+            }
+            else {
+                repo.deleteCommentById(comment_id);
+            }
         }catch (Exception err) {
             err.printStackTrace();
             return "failed";
@@ -72,6 +82,7 @@ public class CommentServiceImpl implements CommentService {
         return "success";
     }
 
+    @Transactional
     @Override
     public String modifyComment(Long comment_id, CommentRequest request) {
         try{
